Products and orders

+++ src/action/auth.action.js
@@ -61,9 +61,23 @@ export const isUserLoggedIn = () => {
 
 export const logout = () => {
     return async dispatch => {
-        localStorage.clear();
         dispatch({
             type: authConstant.LOGOUT_REQUEST,
         });
+        const res = await axios.post('/adminLogout');
+
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({
+                type: authConstant.LOGOUT_SUCCESS,
+            });
+        } else {
+            dispatch({
+                type: authConstant.LOGOUT_FAILURE,
+                payload: { error: res.data.error }
+            });
+        }
+
+
     }
 }
